remove obsolete of GetFile, ExportFile and ListFiles; Add UpdateFileOrFolder
