--- src/appmixer/google/drive/ExportFile/component.json
@@ -1,67 +1,0 @@
-{
-    "name": "appmixer.google.drive.ExportFile",
-    "author": "Jimoh Damilola <jimoh@client.io>",
-    "description": "[OBSOLETE: use DownloadFile instead.] Export file from your Google drive.",
-    "private": false,
-    "webhook": true,
-    "auth": {
-        "service": "appmixer:google",
-        "scope": [
-            "https://www.googleapis.com/auth/drive",
-            "https://www.googleapis.com/auth/drive.appdata",
-            "https://www.googleapis.com/auth/drive.file"
-        ]
-    },
-    "quota": {
-        "manager": "appmixer:google:drive",
-        "resources": "readandwrite",
-        "scope": {
-            "userId": "{{userId}}"
-        }
-    },
-    "inPorts": [
-        {
-            "name": "in",
-            "schema": {
-                "type": "object",
-                "properties": {
-                    "isToken": {
-                        "type": "boolean"
-                    },
-                    "file": {
-                        "oneOf": [
-                            { "type": "string" },
-                            { "type": "object" }
-                        ]
-                    }
-                },
-                "oneOf": [
-                    { "required": [ "isToken" ] },
-                    { "required": [ "file" ] }
-                ]
-            },
-            "inspector": {
-                "inputs": {
-                    "file": {
-                        "type": "googlepicker",
-                        "index": 1,
-                        "label": "File",
-                        "placeholder": "Choose a file...",
-                        "tooltip": "Choose a file to export from Google Drive."
-                    }
-                }
-            }
-        }
-    ],
-    "outPorts": [
-        {
-            "name": "out",
-            "options": [
-                { "label": "File ID", "value": "fileId" },
-                { "label": "File Name", "value": "fileName" },
-                { "label": "Mime Type", "value": "mimeType" }
-            ]
-        }
-    ],
-    "icon": "data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiA/PjwhRE9DVFlQRSBzdmcgIFBVQkxJQyAnLS8vVzNDLy9EVEQgU1ZHIDEuMC8vRU4nICAnaHR0cDovL3d3dy53My5vcmcvVFIvMjAwMS9SRUMtU1ZHLTIwMDEwOTA0L0RURC9zdmcxMC5kdGQnPjxzdmcgZW5hYmxlLWJhY2tncm91bmQ9Im5ldyAwIDAgMzIgMzIiIGhlaWdodD0iMzJweCIgaWQ9IkxheWVyXzEiIHZlcnNpb249IjEuMCIgdmlld0JveD0iMCAwIDMyIDMyIiB3aWR0aD0iMzJweCIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIiB4bWxuczp4bGluaz0iaHR0cDovL3d3dy53My5vcmcvMTk5OS94bGluayI+PGc+PHBvbHlsaW5lIGZpbGw9IiM1MzdBQkQiIHBvaW50cz0iMzEuODY4LDIxIDkuODY4LDIxIDQuODY4LDMwIDI2LjU1NiwzMCAgIi8+PHBvbHlsaW5lIGZpbGw9IiMyRUI2NzIiIHBvaW50cz0iMTAuOTYyLDIgLTAuMDM4LDIwIDQuOTYyLDMwIDE1LjY4MywxMC4zNDUgICIvPjxwb2x5bGluZSBmaWxsPSIjRkVEMTRCIiBwb2ludHM9IjIwLjk2MiwyIDEwLjk2MiwyIDIxLjcxMiwyMSAzMS45NjIsMjEgMjAuOTYyLDIgICIvPjwvZz48Zy8+PGcvPjxnLz48Zy8+PGcvPjxnLz48L3N2Zz4="
-}
